Fix the issues with port mappings in consul version greater than 0.8
Moves the default value generation for ports into the template.
The template uses some jinja2 magic also present in the DebOps project to ensure that the consul_ports dict object is always identical to the config.json file values.
See https://stackoverflow.com/questions/36886650/how-to-add-a-new-entry-into-a-dictionary-object-while-using-jinja2
See https://stackoverflow.com/questions/4870346/can-a-jinja-variables-scope-extend-beyond-in-an-inner-block