Ensure retrieval of existing consul cluster key
Suppose we have a cluster consisting of three consul hosts. The first host has died and has to be newly provisioned.

Setting `run_once: yes` here will run this task only against the facts of that first new host. But that host does not have the encryption key in its config, since it is not configured yet. Therefore, this task will not write anything to the local file system, and the provisioning will silently fail.

Running this task (while delegated to localhost) against all hosts will ensure the `consul_raw.key` will be written to the local file system at least once.