Remove errant 'i' from sed, causing corrupt config
It looks like this was accidentally added when cleaning up this area in
e2f738022648755b727f51c0ef5dc99cd3732a5e. This insert had no effect on
the first run (or if a replication key was statically defined), but on
subsequent runs the replication token would contain this sed pattern and
a newline, causing the JSON prettify-ing of config.json to fail. This
ultimately left the config corrupt and broke other key discovery methods
on further runs.

This is perhaps related to GH issue brianshumate/ansible-consul#160