Fix service deletion in check mode
When installing services on a host for the first time, the role fails in check mode because the list of existing service files is not fetched and thus the `list_current_service_config` fact never gets set.

Initializing it to an empty list in a separate task means it is set properly even if the list of files is empty.