Construct managed_files variable separately on Win
The same task to generate managed_files on Unix hosts can't be used on
Windows because `win_find` is used to generate the list of enabled
services files, and that module returns Windows-style paths with
backslashes. However, `consul_windows_path` is a Unix-style path with
forward-slashes, so we need to use `win_stat` to convert these paths to
be Windows-style paths.

If we don't convert the paths to the correct form, then the task to
delete non-declared services will not match the paths correctly and
all of our JSON files will be deleted.