Introduce role variable: consul_bind_iface
Previously, using this role with hosts that used DHCP for their IP
address could be problematic. When such a host is provisioned, the
config.json file is written with the current IP address in the
consul_bind_address variable. However, if the host reboots and picks up
a new IP address, then the consul service fails when it attempts to bind
to a foreign IP address that doesn't belong to any interface. Note that
0.0.0.0 is also not a valid value for consul_bind_address.

To provide a solution for this, this commit introduces a new variable:
consul_bind_iface. When defined, the `-bind` argument to the service is
instead used with a go-sockaddr template that is resolved at runtime.
For more information, see:

https://www.consul.io/docs/agent/config/cli-flags#address-bind-options

In subsequent commits, related variables will be introduced for similar
consul parameters.