Run ansible-lint in its own workflow
This is in preparation for Molecule 5.x, which removes the `lint`
command. For more information, see:

https://github.com/ansible-community/molecule/pull/3802

Aside from that, this approach is also a bit nicer since linter errors
can be identified by their own GitHub checks run, which should be easier
for developers compared to the previous approach of having to hunt
through Molecule's error logs for linter errors. Also, it's a bit more
efficient since we only run the linters once per CI run instead of in
every Molecule scenario.