Create and delete the tempfile as non-root user
Since the `fetch` task uses non-root privileges, we need to make sure
that this is also the same for handling the tempfile. Otherwise, if this
role is invoked with `become: true` then the non-privileged user won't
be able to `fetch` into the tempfile, since it would be created by root.